Move Search to own file, SearchQuery -> Search
