+++ Sources/App/Commands/Analyzer.swift
@@ -82,7 +82,7 @@ func analyze(application: Application, packages: EventLoopFuture<[Package]>) -> 
     let materializedViewRefresh = statusOps
         .flatMap { RecentPackage.refresh(on: application.db) }
         .flatMap { RecentRelease.refresh(on: application.db) }
-        .flatMap { API.SearchQuery.refresh(on: application.db) }
+        .flatMap { Search.refresh(on: application.db) }
 
     return materializedViewRefresh
 }
