+++ Tests/AppTests/SearchTests.swift
@@ -3,20 +3,20 @@
 import XCTVapor
 
 
-class SearchQueryTests: AppTestCase {
+class SearchTests: AppTestCase {
 
     func test_regexClause() throws {
         XCTAssertEqual(
-            API.SearchQuery.regexClause("foo"),
+            Search.regexClause("foo"),
             "coalesce(package_name) || ' ' || coalesce(summary, '') || ' ' || coalesce(name, '') || ' ' || coalesce(owner, '') ~* 'foo'"
         )
     }
 
     func test_build_single() throws {
         XCTAssertEqual(
-            API.SearchQuery.build(["foo"]),
-            API.SearchQuery.preamble + "\nwhere "
-                + API.SearchQuery.regexClause("foo")
+            Search.build(["foo"]),
+            Search.preamble + "\nwhere "
+                + Search.regexClause("foo")
                 + "\n  order by score desc"
                 + "\n  limit 25"
         )
@@ -24,10 +24,10 @@ class SearchQueryTests: AppTestCase {
 
     func test_build_multiple() throws {
         XCTAssertEqual(
-            API.SearchQuery.build(["foo", "bar"]),
-            API.SearchQuery.preamble + "\nwhere "
-                + API.SearchQuery.regexClause("foo")
-                + "\nand " + API.SearchQuery.regexClause("bar")
+            Search.build(["foo", "bar"]),
+            Search.preamble + "\nwhere "
+                + Search.regexClause("foo")
+                + "\nand " + Search.regexClause("bar")
                 + "\n  order by score desc"
                 + "\n  limit 25"
         )
@@ -46,10 +46,10 @@ class SearchQueryTests: AppTestCase {
                        owner: "owner 2").save(on: app.db).wait()
         try Version(package: p1, reference: .branch("master"), packageName: "Foo").save(on: app.db).wait()
         try Version(package: p2, reference: .branch("master"), packageName: "Bar").save(on: app.db).wait()
-        try API.SearchQuery.refresh(on: app.db).wait()
+        try Search.refresh(on: app.db).wait()
 
         // MUT
-        let res = try API.SearchQuery.run(app.db, ["bar"]).wait()
+        let res = try Search.run(app.db, ["bar"]).wait()
 
         // validation
         XCTAssertEqual(res,
@@ -81,10 +81,10 @@ class SearchQueryTests: AppTestCase {
                        owner: "owner").save(on: app.db).wait()
         try Version(package: p1, reference: .branch("master"), packageName: "Foo").save(on: app.db).wait()
         try Version(package: p2, reference: .branch("master"), packageName: "Bar").save(on: app.db).wait()
-        try API.SearchQuery.refresh(on: app.db).wait()
+        try Search.refresh(on: app.db).wait()
 
         // MUT
-        let res = try API.SearchQuery.run(app.db, ["owner", "bar"]).wait()
+        let res = try Search.run(app.db, ["owner", "bar"]).wait()
 
         // validation
         XCTAssertEqual(res,
@@ -110,7 +110,7 @@ class SearchQueryTests: AppTestCase {
         try packages.map { try Version(package: $0, reference: .branch("default"), packageName: "foo") }
             .save(on: app.db)
             .wait()
-        try API.SearchQuery.refresh(on: app.db).wait()
+        try Search.refresh(on: app.db).wait()
 
         // MUT
         let res = try API.search(database: app.db, query: "foo").wait()
@@ -131,7 +131,7 @@ class SearchQueryTests: AppTestCase {
         try packages.map { try Version(package: $0, reference: .branch("default"), packageName: "foo") }
             .save(on: app.db)
             .wait()
-        try API.SearchQuery.refresh(on: app.db).wait()
+        try Search.refresh(on: app.db).wait()
 
         // MUT
         let res = try API.search(database: app.db, query: "foo").wait()
@@ -149,10 +149,10 @@ class SearchQueryTests: AppTestCase {
             try Repository(package: p, summary: "\($0)", defaultBranch: "master").save(on: app.db).wait()
             try Version(package: p, reference: .branch("master"), packageName: "Foo").save(on: app.db).wait()
         }
-        try API.SearchQuery.refresh(on: app.db).wait()
+        try Search.refresh(on: app.db).wait()
 
         // MUT
-        let res = try API.SearchQuery.run(app.db, ["foo"]).wait()
+        let res = try Search.run(app.db, ["foo"]).wait()
 
         // validation
         XCTAssertEqual(res.results.count, 10)
